+++ apks/ffplayer-1.0/apk/debug/output-metadata.json
@@ -12,7 +12,7 @@
       "filters": [],
       "versionCode": 1,
       "versionName": "1.0",
-      "outputFile": "debug_ffplayer_1.0_1_2021-08-25_14-34-59.apk"
+      "outputFile": "debug_ffplayer_1.0_1_2022-01-04_12-02-01.apk"
     }
   ]
 }
