+++ lib_ffmpeg/src/main/cpp/player/include/MediaDecoder.h
@@ -9,6 +9,7 @@
 #include <android/native_window.h>
 #include <media/NdkMediaCodec.h>
 #include <media/NdkMediaFormat.h>
+
 #include <PlayerResult.h>
