+++ module_recorder/src/main/java/com/pcyfox/module_recoder/ui/activity/LocalVideoPlayTestActivity.kt
@@ -16,8 +16,8 @@ import java.io.File
 
 class LocalVideoPlayTestActivity : AppCompatActivity() {
     private val TAG = "TestActivity"
-    private val url = "/storage/emulated/0/test/20210531_18_30_48/1/video.h264"
-    //private val url = "/storage/emulated/0/test.mp4"
+    //private val url = "/storage/emulated/0/test/20210531_18_30_48/1/video.h264"
+    private val url = "/storage/emulated/0/test.mp4"
     private val storeDir = Environment.getExternalStorageDirectory().absolutePath + "/test/"
 
     private var recordCount = 0
